+++ src/constants/Command.ts
@@ -3,6 +3,8 @@ export const EXTENSION_NAME = "demo-time";
 export const COMMAND = {
   // Documentation
   documentation: `${EXTENSION_NAME}.docs`,
+  // Demo file actions
+  addStepToDemo: `${EXTENSION_NAME}.addStepToDemo`,
   // Running the demo
   start: `${EXTENSION_NAME}.start`,
   previous: `${EXTENSION_NAME}.previous`,